Allow test decorators to also be used as context managers
Using the @contextlib.contextmanager decorator, we can rewrite our
decorators to be generators (using the `yield` statement`), and have
them be usable as either decorators or context managers. This greatly
simplifies the writing of our test utilities, as it eliminates all of
the function-wrapping boilerplate that was otherwise required to create
a custom decorator.