Fix @use_env() decorator causing test to always pass
I think this is because I was using the `yield` and `return` keywords in
the same function. And, according to the Python documentation, a
function with a `yield` keyword makes the function automatically a
generator, even if the `yield` is never executed. According to the
Python documentation for the `yield` statement
(<https://docs.python.org/3/reference/simple_stmts.html#yield>):

"Yield expressions and statements are only used when defining a
generator function, and are only used in the body of the generator
function. Using yield in a function definition is sufficient to cause
that definition to create a generator function instead of a normal
function."